Moved CLI Class one layer up
also fixes the longopt of the report location
